Make Point and Dimension distinct type
While their implementations are identical, the Point and Dimension are
different semantically. When they were just an aliases they could be
interchanged without the compiler ever warning, but since they're now
properly disticint types this error source goes away.

In fact, the immediate effect is that a lot of code didn't compile
anymore, because the signature or arguments were inconsistent.

The property test generators too assumed that point and dimension were
the same type - these are now implemeted as two distinct generators.