Add cubecoords and strongly typed points+dims
Add the cubecoords class to mainatain and sort out mapping between
reference systems. Using this interface should hopefully be less clusmy
than the powerful, but inconvenient local-to-global family of functions.

With this, it became clear that its useful with strongly-typed points
and dimensions, so that coordinates can be combined in a clear way. This
helps documentation with signatures now being a strong hint to how
things work, and to underline the relationships between the coordinate
types.

Mixing reference systems is a very easy mistake to make, and this should
limit at least some of these errors.