Divorce scan/segmenter from openvds
Break away from producing openvds-SEGYScan compatible files, both in
keys and content. This is driven by considerations:

1. The openvds names are quite clunky
2. The contents are not actually openvds compatible, so it's a good
   thing to not have them easily confused

The major content changes is guid being sha256 based, instead of a
timestamp (which, granted, was always the case for this implementation).
Other than that, ranges are no longer closed intervals with begin/end
line numbers - instead, we preserve the full in/crossline table for all
dimensions.

In this iteration, only the three dimensions inline/crossline/samples
are supported, but the format itself is easily generalisable to N
dimensions. In fact, the dimensions themselves are not actually named,
rather indexed in order.