Derive fragment IDs from fragment-grid, not global
Instead of identifying fragments by their top-left coordinate in the
global grid, identify them by the position in the coarsened
fragment-grid.

The main considerations for this decision is that its somewhat simpler
(information is local to the fragment layer only, no leakage), and that
it's trivial to determine neighbour/distance relationships.