Use sha1 for cube IDs
The scheme is to be one-container-per-cube, because permissions, access
etc. are configurable on container level, not so much blob or sub-path
level. However, containers have a max name length of 63 [1], and
stringified sha256 is 64 characters.

sha1 should provide no reasonable vector for collision attacks, which in
turn are not that harmful, so sha1 should do just fine as an object ID.

[1] https://docs.microsoft.com/en-us/rest/api/storageservices/naming-and-referencing-containers--blobs--and-metadata