Add base64 encode/decode
Blob GETs need to be signed with a shared key, which in turns needs some
base64 encode/decode. This is a solved problem [1], so use a pre-written
solution.

Surpisingly, we couldn't immediately find a small, largely-available
base64 encode/decode library. Since we don't want to be in the business
of installing third-party libraries, and base64 is a tiny internal
detail anyway, just vendor it in the source tree, and build as a part of
sc.

[1] https://renenyffenegger.ch/notes/development/Base64/Encoding-and-decoding-base-64-with-cpp