Upload first-dimension slice
The first dimension of the dst view of the source SEG-Y is already
indexed into, and doesn't vary with the x,y,z triple like the rest of
the loop. This is an oversight, but not detected immediately as blob
store is happy to accept empty files (and rightly so).

The first dimension should always be traversed fully.