Add fetch module and libcurl dependency
A fairly simple C++ification of the libcurl interface, mainly dealing
with resources, exception safety and enqueueing of operations.

libcurl requires a substantial bit of code, to spawn and clean up
handles, check for errors. However, in the overall architecture, libcurl
and downloader is seen as a background process/stream/queue that
produces futures for later interaction, namely the extraction of
surfaces and slices from the fragments.

This thing layer centralises scheduling, max-connections, curl
setup/teardown, and lifetime of various handles into a module, to keep
leaks to a minimum.