Rename and revocabularise cubecoords -> gvt
Rename cubecoords, which was a terrible, nonsensical name, to gvt. While
gvt is cryptic on its own, it's a central and easier-to-say name for the
central general volume translation.

With that some docs are polished and added, and some vocabulary cleared
up. This makes the patch somewhat large and messy, but it's reasonable
to consider it one large change and departure.

In effect, the concept of a padded cube is dropped, and the cube is just
the actual source data. Internally, the gvt knows that the cube is
padded for simpler lookup inside fragments, but this is information no
longer leaks.

Method renames
==============
frag_dimension/cube_dimension
-----------------------------
slice_size -> slice_samples

gvt
---
slice_size -> fragment_count
size -> cube_shape, fragment_shape