Rename seismic-cloud.hpp -> geometry.hpp
The role of the previously-named seismic-cloud module was never really
clear cut, but now it's getting more obvious that it is and that there
needs to be a separate geometry module, that deals with the mapping
between cubes, fragments, and coordinate systems. Since all that was
left in the seismic-cloud header after the url and fetch modules is
geometry code, the new name is obvious.