Rename *_point, *_dimension to acronyms CP/CS etc
The *_point and friends were clumsy, and did not in practice convey the
intent well enough to help the total understanding of programs. Instead,
replace these with acronyms which are on purpose quite non-sensical, but
work as mnemonics and *symbols*. It is quite possible that this scheme
will effectively obfuscate code, in which case this change should be
replaced with an even better naming scheme.

With the new names, which are symmetric in length and quite similar in
shape, there follows some new docs to underline their relationship, and
describe the pattern.