Make dimension<ND> take any type, and check domain
Beef up the dimension type to take any type - literals are often used,
and it's also easy to pass it a negative value that is silently
converted to size_t, which gives quite surprising results. By
templating the constructor and making the aware that the source type can
be negative, these errors should be caught. In addition, it means it now
also *verifies* that dimensions are within the domain 0 <= d < ND, so
users can pass unchecked data to the dimension constructor, and let it
handle the error.

With that, the constructor is made explicit so there is no silent
conversion going on. Since a lot of dimension use too boil down to
switching over the possible dimensions, add an implicit conversion into
size_t.