Rename stride -> slice_layout and document
The stride struct and the member names were always a bit rushed, and
unfit for the larger purpose. It is now clearer that it really a larger
layout description for fragments in a cube, and how they translate to
fragments in isolation.

To make the parameters to loops clearer, the members are renamed to
reflect what aspect of the fragment they describe (and, by extension,
how they modify loop parameters), and described in more docs.