Add gvt.slice_stride
This is the cube-based dual to FS.slice_stride, which describes the
layout of a fragment slice in a larger cube. This enables putting
already-extracted slices at the right place in cube-sized slices, quite
simply by having a gvt object with the sliced-dimension set to 1.

The tests are somewhat rough, and the docs surely need some work. A few
illustrations are marked with TODO, and should help aid the total
understanding of what this does significantly.