Build protobuf+grpc in core
The message and service definitions live in the shared protos/
directory, and must be translated to C++ code. Add find_package,
custom_command, and targets to auto-generate and compile protobuf and
grpc source files at build time.

Because of the lack of consistent cmake support in grpc, a fallback code
path using pkgconfig is implemented.

Travis is now required to use bionic, as grpc is unavailable in the
xenial package tree.