Make proto-pkg OneSeismic, Request -> Slice
The protobuf definitions have long been minimum. Now, it's clearer that
the request needs to bring fragment-size information, and is
specifically Slice oriented.

The slice is the first service provided, being the simplest non-trivial
request (in terms of work). To open up the naming and focus the
interface a bit more, the word Slice is introduced onto it.