Use gmtime_r for getting timestamps thread-safely
The gmtime function is not thread safe, with its reliance on a static
buffer. The url functionality will almost always be used from a
multi-threaded context, and should be thread safe.

This puts a dependence on POSIX, since gmtime_r is not standardised
until maybe C2X, and not a part of C++. The pure C++ version of this is
quite clumsy too, and the target is POSIX platforms anyway, so it should
be fine.