Create new curl multi-handle per request
Create a new multi handle for each incoming request, as the curl handles
do not have any internal synchronisation, and the grpc server uses a
thread pool.

This means the server is unlikely to re-use connections between
requests, but that's ok at this point, as it allows the experimentation
on a working, concurrent, albeit potentially slightly slower, server.

This approach is likely to be too expensive for an actual production
implementation, but the design is still not set in stone, and it is more
important that actual end-usage can be tested as well.