Add new curl/transfer worker
The sc::fetch system turned out to be quite impractical after some
experience, and it is more desirable with a simpler, single-task
blocking interface, with callbacks instead.

The one::transfer is an implementation of this - a blocking job that can
internally manage multiple concurrent connections. The main difference
is that it relies on a configuration object with callbacks for
customization points, rather than returning futures. As an overall
component, it doubles down on using curl for transferring files.

The new transfer worker is put in the one:: namespace, with it's header
in the include/oneseismic directory. The project has been named such
for a while, but the old name has stuck around to avoid breaking too
much, and have too noisy diffs. It is expected that everything should be
moved into the one namespace eventually.