Split transfer customisation point
Some experience shows that the transfer configuration is better
partitioned into two objects, the storage configuration, and the
transfer configuration (the action).

The storage configuration is largely static, and tightly bound to the
backing storage (azure, aws, local file system). This should not change
much for the same storage, as it really only deals with fetching
fragments. It is likely this will become a template parameter in the
future, but for now hold onto the dynamic nature of it.

The transfer configuration is the dynamic part - this is expected to be
different for every class of request [1], and is less dependent on
storage.

[1] slice, horizon, single fragments, single traces etc.