Move geometry to the one namespace
Include the geometry module into the include/oneseismic/ directory, and
introduce it in the one namespace. It's sufficiently mature to be
carried into the new codebase.

The actual implementation is just copied and compiled twice, in order
not to break any binaries or up-coming merges - however, everything in
namespace sc is scheduled for removal.