Add worker module + ZMQ dependency
The worker module is responsible for interpreting a (partitioned) job,
fetching and post-processing [1] fragments from storage, and forwarding
partial results back to the job server/session.

The current implementation is a little rough, and is expected to change
substantially with more experience, but should serve as a useful for
testing the system as a whole.

The core library is now made aware of the message format, and has to
build the generated proto library as well.

The worker module is coordinated with ZMQ [2] as transport and
synchronisation, meaning ZMQ is now a dependency.

[1] e.g. slicing fragments and storing the extracted tile
[2] https://zeromq.org/