Add task for fetching cube manifests
This is a rough outline implementation for a task that:
  1. Gets a message
  2. Fetches a manifest from blob
  3. Parses it and formats downstream requests
  4. Pushes these requests onto the next queue

Eventually, this might get the additional responsibility of deciding
exactly how to partition jobs for downstream, but for now it only
produces a single job for a single node.

The sockets are meant to be configured and connected from the outside,
so the task itself is oblivious to network topology, transports etc.
This makes testing a lot easier.

Notably missing:
  - Solid error reporting, this just formats and writes to stderr
  - Error channels to send coordinating messages to the rest of the
    network

Especially the latter makes it slightly more difficult to write tests on
how bad inputs are handled. Additionally, now most bad input or
behaviour will just drop the message (and log), with no up-the-stack
error reporting.