Tie worker to fragment_task instance, add test
The worker function that went into a blocking, infinite loop (save
errors) seemed like a good idea at the time, but has a lot of issues
when it comes to testing, and it takes away some useful flexibility at
call site.

This patch rewrites it to be similar to the manifest-task, and moves the
declaration into the oneseismic/tasks.hpp header, and adds a simple test
that demonstrates a message being pushed through.

Functionally, with the exception of the loop, it should still be
identical.