Replace broken core log format-specifier with UTC
The fmtlib chrono formatters expect a std::tm reference, not a chrono
time point, so the log was always broken. This relies on the POSIX API
for obtaining the tm, as the standard one is not thread safe.

%F %T (ISO 8601) might not be ideal long term, but it is human-readable,
relatable, and comparable across nodes, so it should still provide a
nice starting point for a logging format.