Move http code into .oncomplete, remove .onfailure
The transfer code was written very specifically targeting HTTP, but
both for testing and local filesystem targets, arbitrary (mostly
file://) transport protocols would be quite useful. The status code is a
very http/ftp/smtp specific thing, and more importantly, it is not
always guaranteed that 200/OK means success.

Instead of making a decision on the status code in the libcurl
wrangling, pass the status code to transfer_configuration.oncomplete(),
and remove .onfailure. In the current implementation, on an actual
failed transfer, the code throws an exception, so oncomplete() should
still only be called when the transfer is, on some level, successful.
It's up to the transfer configuration to determine what exactly to do
with the result. This should make file:// based transfer implementable,
help support weird HTTP endpoints, and give more power to the
configuration objects.