Add onstatus(), move http_code out of oncomplete()
Summary
 - Added transfer_configuration::aborted exception class
 - Added transfer_configuration::onstatus
 - Removed status code from transfer_configuration::oncomplete

Remove the status code argument from oncomplete, and add a separate
method onstatus that checks the status code from curl, and decides what
to do next.

This means configurations can be mixed-and-matched a bit more. The main
driver is that what to do on a successful transfer of a fragment is
pretty much the same regardless of storage and transport mechanism
(azure, http://, file:// etc), but how to handle transport details and
even classes of errors is highly dependent on both what's being
transported as well as the storage solution itself.

This makes some new communication necessary - onstatus can decide to
either abort the transfer (by throwing new the aborted exception),
consider it done and continue [1], or retry it.

With it came the realisation that transfer::perform needs to be
exception safe, so to handle it, a function-try-catch is added with
connections reset.

[1] If a config wants to drop certain jobs, returning done would also
    accomplish this