Move onstatus() into storage configuration
Briefly, move the onstatus() method to the storage configuration.

onstatus() being a part of transfer_configuration lead to a pretty
awkward desgin, and introduced a difficult type-and-relationship
problem.

On the actual responsibility bit I find it split pretty much 50/50
whether or not knowledge of the status code belongs with the storage, or
with the transfer. In practice, manifest and fragment fetches will have
different storage configs (for different URL generations), which gives
makes storage ownership a better choice.

On the type issue it is the problem that onstatus() and oncomplete()
need are often overriden [1], but at different levels. The
specific transfer-configuration for a task must somehow expose the
result of a transfer, and this is independent of the storage, but highly
dependent on the task [2]. onstatus() however depends largely on the
storage, but is not so task sensitive. Surely file-not-found will be a
different error if it is the manifest that is wrong (bad input) and the
fragment that's wrong (bad internal state), but that should be handled
by the caller anyway.

Some drive-by changes include removing the az_transfer_configuration,
moving the onstatus() azure default into az, and moving the aborted
exception out of transfer_configuration and into the one namespace.

--

For reference, the type issue comes from

az_transfer_configuration <: transfer_configuration

where az_ implements the onstatus(). The problem is what happens after
oncomplet() and the transfer is completed - how to get the result? No
virtual function [3] is sufficient because of different return types.

A custom fragment_transfer <: transfer_confiugration could work, but
then there is no way of also being az_ without immediately getting the
diamond problem. transfer_configuration is a pure interface now, but it
is quite easy to make it not so, and it's better to avoid the problem
altogether.

--

[1] they are customisation points, after all
[2] fetch-manifest, slice, subcube, trace, horizon etc.
[3] unless returning already-marshalled message, ready for wire