Make azure canonicalized resource overridable
In order to sign a request, azure expects the canonicalized resource as
a part of the message to be signed. This will have a different format
when signing the manifest (as it has a different URL scheme), but other
than that it could use the same sign code. The default implementation
is unchanged, and works for fragments only.

This is more a temporary fix than anything. Inheritance for code reuse
is awful, and the manifest code reads like an afterthought (which it
really is). A more sustainable change would be to make this actually
dynamic with format-strings, or properly separate implementations. This
will be a task for the future.