Add fail-channel to manifest-task for notfound
Add a failure channel to the manifest_task::run() to signal both soft
and hard failures to other nodes.

Soft failures means things like bad user input - if a non-existing
manifest is requested, permissions are denied, or similar, it is a soft
error that should be logged, and the corresponding error message should
bubble back to the caller.

Hard errors are evidence of attacks, or infrastructure (such as nodes
not being available). They should be logged and communicated similarly,
but given different severity levels and topics.

The notfound exception is added. It is still a bit in the air whether or
not HTTP codes like 404 and 500 should raise C++ exceptions, or be woven
into return codes. The main issue arises from file system addressing
this differently, and it being somewhat awkward to communicate from
inside transfer. For now, the exception seems like a reasonable way of
signaling it [1], and regardless a larger interface change is needed to
use return codes.

[1] Even more so if it is assumed requests should be pre-vetted