Signal manifest-task failures with exceptions
Redesign the manifest-task use cached objects, and to signalise
operation failures [1] with exceptions. This makes the control flow of
the run() function a lot clearer and high-level, with low level details
pushed into helper functions.

I experimented a lot to arrive at this, and tried:
 - Returning an enumerator
 - Exceptions, but try {} catch {} around the expression that would
   throw it
 - Return codes from helpers
 - Everything inline in run()

This was the design that gave the clearest and most obvious code. The
largest difference from most [2] C++ exception use is using exceptions
for common and expected failures from user input, e.g. non-existing
cubes.

A request for a non-existing cube can be considered just bad user input,
but the only way [3] to determine if a cube is missing is trying to
fetch it, and the input was *structured* correctly. Regardless, from the
point of view of run(), the procedure must be aborted, and the fact that
the job *is* aborted must be signaled back to the session manager, and
is at that point indistinguishable from a lot of other internal
failures, and data integrity and service availability is not
compromised.

[1] both bad user input and connection issues are considered failures
[2] I have nothing but anecdotal data for backing this up
[3] At least currently