Clarify not-found failure tests with multiparts
Use the cppzmq multipart message interface to clarify the intent of the
tests, and emphasise the protocol of the failure channel, with using
multipart messages.

The use of the word 'envelope' always suggested it was arbitrary data
that should be pushed through, when in fact there is a protocol for
failure signaling too.