Move did-socket-recv-msg test to shared utility
Introduce a test-utility file with a function for checking if there are
messages waiting on a socket.

Verifying that a failure scenario did not send a message is necessary
for (maybe?) all tests that touch on the topic, and the correct
implementation [1] is somewhat clunky. Additionally, a good name for
this makes the test so much more obvious than randomly doing
recv(dontwait) on a socket.

[1] with poll