Use protocol helpers in multiplexer test
Instead of a dumb piping of bytes, use the partialResult and
partitionRequest message types to give some structure to the bytes. This
makes the actual behaviour of the goroutine easier to follow, as it
instead of being plumbing actually pretends to work as the components
it is meant to emulate.