Split partialResult in a base and routed part
Add the routedPartialResult and set the address only there. The return
address will have to flow through most of the system, but will be
consumed by the ZMQ routing in the recv() part of multiplexing. Having a
composed struct can remove the need for a passing raw byte arrays
around, and is a nice peg hang documentation onto.