Return error from loadZMQ
Catching unexpected input early really helps keeping the system in a
valid state, and the translation from bytes to structs is the first
point of contact of unsanitized data. Add an error type to the loadZMQ
so that multipart messages with an unexpected number of parts are
discovered and logged.