Terminate process on loadZMQ errors
Check the error of loadZMQ and terminate the process if the message
layout is not as expected. This is certainly out of the question for an
actually operational service, but the infrastructure for recovering and
logging is not in place, and it is better to terminate development
deploys than to infinitely block or corrupt data.

While sprinkling log.Fatal all over the place is quite bad style, it
does give quick termination at the first detection of failure. This is
very noticable when tinkering with the message format for testing
purposes, since the tests will just hang if fewer messages are sent than
what's expected.

It is quite clear that all these fatals must go away pretty soon, but
they should be easy to find and replace as they stand out.