Rename partitionRequest -> process
The partitionRequest name was never really great, as it pulled focus
away from what's happening in the session module, by mudding it with
internal details from another module.

The process is a new name (lifted from unix) that should be used all
over the code for the same concept, namely the macro view from a
request hits the API server until it is fully served.