Replace multiplexer with sessions object
Replace the raw multiplexer with a sessions object, responsible for
managing sessions, creating channels, and routing messages to/from ZMQ
queues and dependent channels.

The multiplexer function always was always the workhorse that arranged
all this, but slightly underhanded in doing so. There will be a need to
track live requests across the system, cancel them when they fail, and
route results back to the right user connection, and so more formalism
needs to be in place.

This turned out to bring out quite a lot of various small changes. I'd
like to keep it even smaller, but in doing so the changes became
non-sensical.