Explicitly build multiparts in plan/fetch phase
Instead of just replacing the body, explicitly build the messages. This
makes the format clearer in code, and keeps it more similar to the go
implementation.

This actually demonstrated a pretty significant bug - the address was
interpreted as the pid in the C++ code. This is fixed, and to make it
clearer some less noisy ZMQ interfaces are used in test code to keep
more emphasis on the message format.