Set query/fragment IDs outside helper functions
Partitioning a request boils down to distributing the fragment IDs, as
all worker nodes get a full copy of the rest of the request. While the
slice() and family of helper functions will be useful for
setting request-specific query parameters, there's no reason to do so
just to populate the fragment IDs field. Furthermore, since breaking up
the fragment IDs is independent of how the fragments are then
re-assembled, it makes perfect sense to use the same paritioning code
for all requests.