Add protocol support for partitioned requests
Extend the protocol between the manifest/planner, the fragment/worker,
and the session/api server modules to include a message for n/m.

This patch got a bit larger than I'd like, but protocol changes are
awfully invasive. Maybe this is evidence of them being too invasive, and
that work should also go into making fewer protocol details leak into so
much code, maybe in test code (which take the focus away from what's
being tested).

That aside, what this change really boils down to is adding a string
"n/m" (read: n of m), where n is zero-based, and m is the number of
tasks in this process. Having n as an index and m as cardinality makes
implementions simple - allocating space for a bitstring to track
completed tasks can be done by just parsing m, and what bit to set can
just use n as an index.

While the n/m information is added with this patch, partitioning and
assembly is not implemented. While not very much code is needed to
add partitioning, there's always the risk for bugs that will be easier
to pinpoint with this being in two commits.

"n/m" as a string feels somewhat awkward, and is likely bad for
performance, but it was reasonably easy to implement, and should serve
well while getting some experience with partitioning. Parsing integers
from bytes adds some complexity (endianness for one) and was quite
awkward in go. Additionally, the internal interfaces are not yet good
enough to ergonomically work with this slightly more sophisticated
protocol, so keeping the ad-hoc parsing simpler makes sense, until some
refactoring can be done.