Create multiple fetch tasks per process
Split the set of fragment IDs into multiple separate tasks, to allow
fetch to run in parallel across multiple nodes. The task size is set a
priori, and is not dynamic, as it is not obvious that dynamic task
scalign will even be beneficial, or at least be worth it in terms of
complexity.

The resulting fragments are buffered and put in order at the session/api
node, so clients are unaware of the (possible) parallelism in task
fetch. This means unmodified clients will still work as expected.

Some error and failure signaling is glaringly missing, but as of now
there are no obvious places to signal and fail processes (particularly
in go). This is not made very much worse by parallel fragment tasks, so
it is acceptable for this change.