Replace start_processing() with parse()
The start_processing() function only parsed the pid, but the new parse
function parses the full message protocol, to have a single point in
code where the message is translated from bytes to its semantics.

The start-processing concept can be reintroduced later, but did not pay
off now. This change makes understanding the message format a lot
easier, and the message format is probably the single largest complexity
factor.

Notably, the process name is now introduced properly in the manifest
module.