Test that HTTP headers are set
Add a test for checking that HTTP headers are correctly set on the
HTTP request performed by curl. While this borders on checking that curl
works, it's quite easy to make programming mistakes when overriding
http_headers() or by returning values in a surprising format, so having
a basic confidence that correctly-added headers are actually sent to the
HTTP server is a useful safety net.