Use behalf-of authorization for all requests
The use of a pre-shared key for authorizing requests to Azure is
deprecated in favour of bearer tokens. The immediate API was already
aware of this change, but requests to azure were still performed with
pre-shared keys.

For this to work, the authorization header must become a part of the
messages, and must flow all the way from outer API request to the HTTP
requests handled by libcurl.