Remove key arg from manifest+fragment apps
With the deprecation of pre-shared keys, the --key argument to these
binaries is dead. Please noe that the one::az signature is not changed,
but instead passed an empty string, to keep the change focused on
program setup.

It's likely that all pre-shared key details will be removed from
one::az some other time.