Create socket identity inside Sessions object
Rather than injecting the socket identity, have it be created inside the
newSessions functions, and bind it to the Sessions struct. In effect,
this is not a change to anything, since the the identity was passed as a
parameter to the Run() function which closes over its arguments and goes
into an infinite loop.

Using caller-defined identities is easily re-enabled by creating the
Sessions object manually, or by adding another New() function.

The main benefit is a less cumbersome setup, for something that really
boils down to a message protocol detail anyway.