Move sessions object out of slice controller
Move the sessions object to the outer-most layer of the app, rather than
being an internal goroutine created by the slice controller. With the
sessions setup and run being done in the app init function, the
createSliceController function does not do anything but forward params
to the struct, so init in inline instead.

This means the sessions instance, which HTTP requests to ZMQ messages
and is unaware of the underlying request, can be shared between multiple
controllers.