Signal and cancel processes on failure message
Send a signal on the proIO.err channel that a process has failed,
whenever its pid is read from the ZMQ failure queue. This should close
the out and err channels, and provide enough information to the models
to give a good, user-friendly error message.

Sturcturally, this is pretty similar to how replies are mapped from ZMQ
messages to goroutines already. There are some opportunities for
refactoring that become clearer with this feature.