Close io.out before sending on io.err
Fix the possible race condition when failing a process, by closing the
io.out channel before sending messages on io.err. This seemingly
pointless change allows callers to use range-for on io.out to read from
it, which always was the intention.

The corresponding test is updated to reflect this, which addresses two
problems:

1. It demonstrates the intended pattern for consuming the channels,
   i.e. range-for on both out and err, or m, _ = <-err
2. It fixes the issue of the io.out case being select'd when the channel
   is closed, even when no message is waiting.