Drop messages from already-cancelled processes
Drop (and log) messages belonging to processes that have already been
cancelled, i.e. the entry is removed from the process table. Testing it
is a bit iffy with the current design, but I managed to quite reliably
get it to crash on nil access without the !ok check.

Messages coming in after the process has been terminated should be a
fairly common event.