Return HTTP status other than 404 in slice GET
When a slice GET request fails, return more precise error codes than
just always 404 Not found, by checking the io.err channel.

The new failure error type maps between the failure identifiers, encoded
as strings, and HTTP status code and error messages. It's a fairly
simple and straight forward approach, but gives little room as-is for
richer failure reporting. newFailure() should provide a decent starting
point for more sophisticated interpretation of messages off the failure
queue, should it be necessary.

The failure implements the error interface so it can be returned along
with other errors that aren't pullled from the failure queue. This means
the signature for fetchSlice is unchanged, but some slightly ugly type
selecting code is needed to determine what to do with the error.