Replace gvt::slice_stride -> gvt::injection_stride
slice_stride was used to compute the stride for inserting a ND-1 slice
into an ND cube, usually setting the slice dimension to 1, using
gvt<ND>. This further required the FID to e set to zero along the slice
dimension.

Replacing this with injection_stride, a function that computes the
stride for inserting an ND fragment int on ND cube, and rather removing
a dimension on gvt feels a bit more intuitive and declarative.