Do not upgrade packages when running go generate
go.mod and go.sum should not be considered artifacts.
The swag and protoc-gen-go commands need to be installed during build.
Their versions should respect go.mod and go.sum.
The "go get" command gets the last version of the package and updates
go.mod and go.sum.
The "go install" command seem to behave as we need.