Make single oneseismic python package
Unify the scan and upload packages into a single oneseismic python
package.

Move all the source code from the ingestion/scan/, ingestion/upload/,
and client/ dirs into python/. These modules are functionally
independent, but for distribution and naming purposes it is reasonable
to keep them in the same module, or at least in the same namespace. This
means a single readme, a single root for docs, a single entry point and
a single name.  Naturally, that also means a single version, but while
the programs can be run and written independently, most significant
changes (to in/output formats in particular) would have to happen in
lockstep anyway.

The oneseismic package does not quite implement a package namespace [1].
While this is certainly something to consider should the need arise, it
would mean more packaging files [2] and possibly a less ergonomic
__main__.py.

The main features brought on by this are:
- fewer directories
- a single package upload to pypi
- a single readme
- a single setup.py
- a single package program entry point

The final point is really the ability to run scan and upload with the
same executable:

  python3 -m oneseismic scan args...
  python3 -m oneseismic upload args...
  python3 -m oneseismic login args...

Any new program can be added to this set of sub commands.

The single package approach has drawbacks, but should provide a nicer
user experience, and a nicer developer experience to boot [3].

[1] https://packaging.python.org/guides/packaging-namespace-packages/
[2] setup.py, setup.cfg, pyproject.toml
[3] since everything is in the same repository