Close error channel before output can be consumed
This is a simplification. Also sending the error to the client after
starting to send the data is too complex for now.

The status code of the response can only be set once.

Keep track of the number of fragments to be sent. Then the client can
know if all fragments have been received, and inspect the log if needed.