Send each of the fragments as a separate json
Parsing the entire response to json uses too much memory. Might be poor
implementation, but json is not meant for this large datasets.

Each fragment is manually embedded in a json list.

If the transfer is closed before all fragments have been received the
json list will not be closed and will be invalid.