Rewrite messaging and API servers
This commit is unfortunately *massive*, but during it became clear quite
early that in order to address the issues that I set out to do,
everything would break as a result. I think that trying to break it up
into smaller pieces would mean too many changes and too just to keep
code scheduled for removal running, and not worth the development cost
nor log complexity.

Key changes
-----------
- Protobuf is out, in favour of json and messagepack
- Single-direction message flow
- Servers re-written, more tests added
- Gin-gonic replaces iris as http server framework

Goals
-----
The main goal of this patch is to simplify message flow, and instead of
providing a synchronous request-response API, create a new resource for
the result which can be queried at a later time. This is somewhat
chicken-and-egg - not having to route data back to a living, current
request reduces the need for more complex messaging, and makes every
step a pipeline. Then, messages are simplified because they don't need
return addresses etc.

The application was so deeply involved in the messaging patterns and
implementation details of protobuf that changes cascaded very fast.

This is a rough sketch of message flow *before* this patch:

[user]
-> [api/main] <------------------------\
-> [manifest] <-> [blob/manifest.json] |
        \_______<-____/                |
-> \                                   |
   |+ [fragment/worker 1] -> +         |
   |+ [fragment/worker 2] -> +         |
   |+ [fragment/worker 3] -> +         |
                             \---->---/

And after:

[user]
-> [api/query]
-> [manifest] -> [blob/manifest.json]
        \_______<-____/
-> \
   |+ [fragment/worker 1]
   |+ [fragment/worker 2]
   |+ [fragment/worker 3]
-> [relay]
-> [blob/results]

[user]
<-> [api/result] <-> [blob/result.f32]

where blob/ in both cases means blob store. The client is required to
make two requests - one to start the process, and one to fetch the
now-prepared result.

On protobuf and messages
------------------------
Protobuf has been replaced by a combination of json and messagepack [1],
for a few key reasons:

1. Protobuf is a black box, and is concatenating messages without
   serializing them is infeasible. This is the key driver behind the
   move to messagepack.
2. Protobuf is nice in the sense that it has a schema, but the auto
   generation of code, while it makes it available for more languages,
   painful.
3. Protobuf autogenerated code is quite unergonomic, which made all the
   code touching it quite inelegant. The intent of messages, and to a
   lesser degree the messages themselves, were drowned in noise.
4. The messages would typically have to be mapped to a proper
   language-native type anyway, and this is as much work with protobuf than
   it is with any other serialization library.
5. Compiling with protobuf is more painful than the alternatives.

Messagepack is used whenever there's binary data (floats) involved, and
JSON for task descriptions.

This is not without its other benefits. Since message types now always
come with a language-native representation (e.g.
include/oneseismic/messages.hpp), it's much more natural to test. They
all come with pack+unpack methods and do not in code specify what
serialization mechanism to use - this means that changing to other
serialization formats in the future should mean little or no change in
calling code. Serialization and verification is now tested, and the
actual on-the-wire formats are more annotated and documented. This is
certainly not made more difficult with protobuf, but for several reasons
never really happened.

On gin-gonic and iris
---------------------
Since the rewrite was so comprehensive in terms of messages, the port
from iris to gin-gonic was thrown into the mix. From before there were
few and quite poor unit tests in the api/ dir, and it was never really
designed to be tested. This made making improvements and changes to
messaging, scheduling and primarily failures difficult.

Iris was abandoned largely because of community worries about upstream
development. Gin-gonic seems well established and measures up pretty
well.

Replacing the http server framework does not really in itself help
achieving the goal of this patch. It started out as accidental, but
a nice opportunity arose to start with a clean slate and fresh
assumptions.

On golang project structure
---------------------------
The api/ dir has been restructured to look like a standard golang
project [1]. It's designed a single module with shared internal
components (auth is the obvious one), and built as different binaries.
Right now the /query and /result call are both served by the same
server process, but this is really only because it means not repeating
the argument parsing and middleware init. In the (near) future, it is
expected that /query and /result will be served by different processes.

The programs can be runtime configured with command line parameters in
addition to environment variables, which brings on a --help option and
much easier configuration for one-off runs.

Summary
-------
This patch brings on a lot of big changes, but even though virtually all
of the infrastructure is rewritten, there are fairly few changes in the
client. Really, it seems to boil down to performing two HTTP requests
instead of one, and importing messagepack in addition to ujson.

It's not obvious yet how internal storage is going to work, nor how jobs
should be (globally) cancelled and failure signalled. For
ease-of-implementation, partial results are written straight to azure,
but there's no problem running it through some aggregation, or storing
in a database of some kind. While oneseismic is nowhere compelete with
this patch, it yanks it towards a more elegant pattern of messaging, and
a better testbed for exploring solutions.

[1] https://msgpack.org/
[2] https://github.com/golang-standards/project-layout